requirements.txt
pandas==2.0.3
plotly==5.15.0
haversine==2.7.0
inflection==0.5.1
numpy==1.24.4
folium==0.14.0
import streamlit as st 1.24.1
streamlit-folium == 0.7.0
Pillow == 9.5.0